Started working on graphics (only rendering primitives) and physics (only sphere-sphere contact detection). Now using quaternions.

+++ R3D/src/R3D/Core/Layer.h
@@ -48,7 +48,7 @@ namespace r3d
 		virtual void onStepBack(Window* window) {};
 		virtual void onStepForward(Window* window) {};
 
-		int getFrameSlotIndex() const { return m_savedFrames.getActiveSlotIndex(); }
+		size_t getFrameSlotIndex() const { return m_savedFrames.getActiveSlotIndex(); }
 
 	protected:
 		std::string            m_name;
